+++ api-webapp/src/main/kotlin/org/jeudego/pairgoth/store/FileStore.kt
@@ -106,6 +106,8 @@ class FileStore(pathStr: String): Store {
     private fun buildTournament(json: Json.Object): Tournament<*> {
         val tournament = Tournament.fromJson(json, canonicalize = false)
         tournament.lastAction = json.getString("lastAction")
+        tournament.lastSync = json.getLong("lastSync")
+        tournament.lastPairing = json.getLong("lastPairing")
         var maxPlayerId = 0
         var maxGameId = 0
         val players = json["players"] as Json.Array? ?: Json.Array()